requirements.txt
streamlit
Pillow
google-generativeai
opencv-python
numpy
